src/requirements.txt
supabase
python-dotenv
APScheduler
